Fix quick start in documentation
Remove irrelevant badge and headers
